tests: fs: Make common fs_basic test
Make test_fs_basic test common.
